added deleteing event to AdminEventService

+++ ticket-manager/event/src/main/java/pl/ticket/event/admin/event/repository/AdminEventRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import pl.ticket.event.admin.event.model.AdminEvent;
 
 @Repository
-public interface AdminEventRepository extends JpaRepository<AdminEvent, Integer>
+public interface AdminEventRepository extends JpaRepository<AdminEvent, Long>
 {
 }
